move component into a new file
